3645 - move common hooks to Dashboard/hooks
