+++ src/client/components/Dashboard/hooks/useGetTableData.tsx
@@ -6,8 +6,7 @@ import { Table } from 'meta/assessment'
 import { useAppDispatch } from 'client/store'
 import { DataActions } from 'client/store/data'
 import { useCountryRouteParams } from 'client/hooks/useRouteParams'
-
-import { useDependencies } from './useDependencies'
+import { useDependencies } from 'client/components/Dashboard/hooks/useDependencies'
 
 export const useGetTableData = (table: Table) => {
   const dispatch = useAppDispatch()
